Split systemVersionPlist from lastBuid
